Updating package scripts.

+++ config/RPM/centos_postinstall.sh
@@ -7,8 +7,19 @@ SHARE=/usr/share/periscope
 LOG=/var/log/blippd.log
 
 useradd -r ${USER} -d ${HOME}
+if [ ! -d ${HOME} ]; then
+    mkdir ${HOME}
+fi
+
+chown ${USER}:${USER} ${HOME}
+
 touch ${LOG}
-chown ${USER} ${LOG}
+chown ${USER}:${USER} ${LOG}
+
 cp ${SHARE}/blippd.conf ${PETC}/
 cp ${SHARE}/blippd.service /etc/systemd/system/
-cp ${SHARE}/blippd.opts /etc/sysconfig/blippd
+
+if [ ! -f /etc/sysconfig/blippd ]; then
+    cp ${SHARE}/blippd.opts /etc/sysconfig/blippd
+fi
+systemctl daemon-reload
